+++ src/main/java/com/multi/mlpenterpriseapprovalsystem/schedule/todo/controller/TodoController.java
@@ -1,0 +1,18 @@
+package com.multi.mlpenterpriseapprovalsystem.schedule.todo.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+/**
+ * 투두리스트 컨트롤러
+ *
+ * @author : 권지영
+ * @filename : TodoController
+ * @since : 2026. 1. 4. 일요일
+ */
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/v1/todo")
+public class TodoController {
+}
