+++ src/main/java/com/multi/mlpenterpriseapprovalsystem/schedule/todo/repository/TodoRepository.java
@@ -1,0 +1,14 @@
+package com.multi.mlpenterpriseapprovalsystem.schedule.todo.repository;
+
+import com.multi.mlpenterpriseapprovalsystem.schedule.todo.domain.TodoList;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+/**
+ * 투두리스트 레포지토리
+ *
+ * @author : 권지영
+ * @filename : TodoRepository
+ * @since : 2026. 1. 4. 일요일
+ */
+public interface TodoRepository extends JpaRepository<TodoList, Long> {
+}
